Fixes circular dependency (#439)
