+++ src/selectors/entities/users.js
@@ -8,8 +8,9 @@ import {
     getCurrentUser,
     getCurrentUserId,
     getUsers,
+    getMyCurrentChannelMembership,
 } from 'selectors/entities/common';
-import {getMyCurrentChannelMembership} from 'selectors/entities/channels';
+
 import {getDirectShowPreferences} from 'selectors/entities/preferences';
 
 import {filterProfilesMatchingTerm, sortByUsername, isSystemAdmin} from 'utils/user_utils';
